+++ mainGndMapRdr/main.cpp
@@ -37,8 +37,8 @@ static class TestStation* sys = 0;
 // updateDataCB() -- Station's background tasks
 static void updateDataCB(int)
 {
-    double dt0 = 1.0/double(bgRate);
-    unsigned int millis = (unsigned int) (dt0 * 1000);
+    double dt0 = 1.0/static_cast<double>(bgRate);
+    unsigned int millis = static_cast<unsigned int>(dt0 * 1000);
     glutTimerFunc(millis, updateDataCB, 1);
 
     // Current time
@@ -124,8 +124,8 @@ int main(int argc, char* argv[])
 // Set timer for background tasks
 // ---
 
-    double dt = 1.0/double(bgRate);
-    unsigned int millis = (unsigned int) (dt * 1000);
+    double dt = 1.0/static_cast<double>(bgRate);
+    unsigned int millis = static_cast<unsigned int>(dt * 1000);
     glutTimerFunc(millis, updateDataCB, 1);
 
 // ---
